requirements.txt
gradio
PyPDF2
pinecone_client
python-dotenv
sentencepiece
tensorflow
tf-keras
cohere